% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsverse-package.R, R/datasetsverse.R
\docType{package}
\name{DataSetsVerse}
\alias{DataSetsVerse}
\alias{DataSetsVerse-package}
\title{DataSetsVerse: A Metapackage for Thematic and Domain-Specific Datasets in R}
\usage{
DataSetsVerse()
}
\value{
Invisibly returns the names of the loaded packages.
}
\description{
{
This metapackage provides a curated collection of R packages containing domain-specific datasets. It includes time series data, educational metrics, crime records, medical datasets, and oncology research data.
}

This function displays a formatted list of the datasets packages included
in the DataSetsVerse metapackage and their respective versions.
}
\details{
DataSetsVerse: A Metapackage for Thematic and Domain-Specific Datasets in R

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A Metapackage for Thematic and Domain-Specific Datasets in R.
}
\seealso{
{
Useful links:
\itemize{
  \item \url{https://github.com/lightbluetitan/datasetsverse}
}
}
}
\author{
{
\strong{Maintainer}: Renzo Caceres Rossi \email{arenzocaceresrossi@gmail.com}
}
}
